refactor: split exception handlers into single-responsibility files
Split monolithic exception handler classes in notification, scheduling,
security, and subscription modules into individual files — one per
exception type. Aligns with the pattern already established in admin,
coaching, identity, and location modules.
