+++ src/main/kotlin/com/nickdferrara/fitify/scheduling/internal/advices/ScheduleConflictExceptionHandler.kt
@@ -1,0 +1,20 @@
+package com.nickdferrara.fitify.scheduling.internal.advices
+
+import com.nickdferrara.fitify.scheduling.internal.controller.ClassController
+import com.nickdferrara.fitify.scheduling.internal.controller.WaitlistController
+import com.nickdferrara.fitify.scheduling.internal.dtos.response.ErrorResponse
+import com.nickdferrara.fitify.scheduling.internal.exceptions.ScheduleConflictException
+import org.springframework.http.HttpStatus
+import org.springframework.http.ResponseEntity
+import org.springframework.web.bind.annotation.ExceptionHandler
+import org.springframework.web.bind.annotation.RestControllerAdvice
+
+@RestControllerAdvice(assignableTypes = [ClassController::class, WaitlistController::class])
+internal class ScheduleConflictExceptionHandler {
+
+    @ExceptionHandler(ScheduleConflictException::class)
+    fun handleScheduleConflict(ex: ScheduleConflictException): ResponseEntity<ErrorResponse> {
+        return ResponseEntity.status(HttpStatus.CONFLICT)
+            .body(ErrorResponse(ex.message ?: "Schedule conflict"))
+    }
+}
